Move StorageProperties class into its own file
